add : bottomsheet filter to component
