+++ lib/screens/main/dashboard/dashboard.dart
@@ -134,8 +134,7 @@ class _DashboardState extends State<Dashboard> {
                     ),
                   ),
                 ),
-
-                const Padding(
+                Padding(
                   padding: EdgeInsets.only(right: 8.0, left: 8.0, top: 130.0),
                   child: SizedBox(
                     height: 155,
@@ -151,7 +150,7 @@ class _DashboardState extends State<Dashboard> {
                 children: [
                   Text(
                     "Rekomendasi Untukmu",
-                    style: TextCollection.bodySmall,
+                    style: textApp.bodySmall,
                   ),
                   const Spacer(),
                   IconButton(
@@ -178,7 +177,7 @@ class _DashboardState extends State<Dashboard> {
                 children: [
                   Text(
                     "Gudang Termurah",
-                    style: TextCollection.bodySmall,
+                    style: textApp.bodySmall,
                   ),
                   const Spacer(),
                   IconButton(
@@ -200,405 +199,6 @@ class _DashboardState extends State<Dashboard> {
           ],
         ),
       ),
-    );
-  }
-
-  Future<dynamic> FilterBottomSheet(BuildContext context) {
-    return showModalBottomSheet(
-      context: context,
-      isScrollControlled: true,
-      backgroundColor: ColorApp().light4,
-      builder: (builder) {
-        return StatefulBuilder(
-          builder: (context, setState) {
-            return Container(
-              height: MediaQuery.of(context).size.height * 0.9,
-              decoration: const BoxDecoration(
-                borderRadius: BorderRadius.vertical(
-                  top: Radius.circular(20),
-                ),
-              ),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Padding(
-                    padding: const EdgeInsets.symmetric(horizontal: 16.0),
-                    child: SizedBox(
-                      height: 50,
-                      child: Row(
-                        children: [
-                          Icon(
-                            Icons.filter_list_rounded,
-                            color: ColorApp().dark1,
-                          ),
-                          const SizedBox(width: 20),
-                          Text("Filter", style: textApp.bodySmall)
-                        ],
-                      ),
-                    ),
-                  ),
-                  const Divider(),
-                  const SizedBox(height: 10),
-                  Padding(
-                    padding: const EdgeInsets.symmetric(horizontal: 16.0),
-                    child: Text("Urutkan", style: textApp.bodySmall),
-                  ),
-                  // const SizedBox(height: 10),
-                  SizedBox(
-                    height: 30,
-                    child: Row(
-                      children: [
-                        Checkbox(value: false, onChanged: (value) {}),
-                        const SizedBox(width: 5),
-                        Text(
-                          "Paling Rekomendasi",
-                          style: textApp.smallLabel.copyWith(
-                            color: ColorApp().dark1,
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                  SizedBox(
-                    height: 30,
-                    child: Row(
-                      children: [
-                        Checkbox(value: false, onChanged: (value) {}),
-                        const SizedBox(width: 5),
-                        Text(
-                          "Harga Terendah",
-                          style: textApp.smallLabel.copyWith(
-                            color: ColorApp().dark1,
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                  SizedBox(
-                    height: 30,
-                    child: Row(
-                      children: [
-                        Checkbox(value: false, onChanged: (value) {}),
-                        const SizedBox(width: 5),
-                        Text(
-                          "Harga Tertinggi",
-                          style: textApp.smallLabel.copyWith(
-                            color: ColorApp().dark1,
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                  const Divider(),
-                  const SizedBox(height: 10),
-                  Padding(
-                    padding: const EdgeInsets.symmetric(horizontal: 16.0),
-                    child: Text("Price Range", style: textApp.bodySmall),
-                  ),
-                  RangeSlider(
-                    min: 0,
-                    max: 100,
-                    activeColor: ColorApp().mainColor,
-                    inactiveColor: ColorApp().mainColor.withOpacity(0.5),
-                    values: _sliderValue,
-                    labels: RangeLabels(
-                      _sliderValue.start.toString(),
-                      _sliderValue.end.toString(),
-                    ),
-                    onChanged: (newValues) {
-                      setState(() {
-                        _sliderValue = newValues;
-                      });
-                    },
-                  ),
-                  Padding(
-                    padding: const EdgeInsets.symmetric(horizontal: 16.0),
-                    child: Row(
-                      children: [
-                        Column(
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: [
-                            Text(
-                              "Min. price",
-                              style: textApp.extraSmallLabel.copyWith(
-                                color: ColorApp().dark1,
-                              ),
-                            ),
-                            SizedBox(
-                              height: 30,
-                              width: 80,
-                              child: TextField(
-                                cursorColor: ColorApp().mainColor,
-                                keyboardType: TextInputType.number,
-                                decoration: InputDecoration(
-                                  focusedBorder: OutlineInputBorder(
-                                      borderRadius: BorderRadius.circular(5),
-                                      borderSide: BorderSide(
-                                        color: ColorApp().mainColor,
-                                      )),
-                                  contentPadding: const EdgeInsets.only(
-                                    bottom: 5.0,
-                                    left: 5.0,
-                                  ),
-                                  border: OutlineInputBorder(
-                                    borderRadius: BorderRadius.circular(5),
-                                    borderSide: BorderSide(
-                                      color: ColorApp().mainColor,
-                                    ),
-                                  ),
-                                ),
-                              ),
-                            ),
-                          ],
-                        ),
-                        const Spacer(),
-                        Column(
-                          crossAxisAlignment: CrossAxisAlignment.end,
-                          children: [
-                            Text(
-                              "Max. price",
-                              style: textApp.extraSmallLabel.copyWith(
-                                color: ColorApp().dark1,
-                              ),
-                            ),
-                            SizedBox(
-                              height: 30,
-                              width: 80,
-                              child: TextField(
-                                cursorColor: ColorApp().mainColor,
-                                keyboardType: TextInputType.number,
-                                decoration: InputDecoration(
-                                  focusedBorder: OutlineInputBorder(
-                                      borderRadius: BorderRadius.circular(5),
-                                      borderSide: BorderSide(
-                                        color: ColorApp().mainColor,
-                                      )),
-                                  contentPadding: const EdgeInsets.only(
-                                    bottom: 5.0,
-                                    left: 5.0,
-                                  ),
-                                  border: OutlineInputBorder(
-                                    borderRadius: BorderRadius.circular(5),
-                                    borderSide: BorderSide(
-                                      color: ColorApp().mainColor,
-                                    ),
-                                  ),
-                                ),
-                              ),
-                            ),
-                          ],
-                        ),
-                      ],
-                    ),
-                  ),
-                  const SizedBox(height: 20),
-                  const Divider(),
-                  Padding(
-                    padding: const EdgeInsets.symmetric(horizontal: 16.0),
-                    child: Text(
-                      "Payment Time",
-                      style: textApp.bodySmall,
-                    ),
-                  ),
-                  // const SizedBox(height: 10),
-                  SizedBox(
-                    height: 30,
-                    child: Row(
-                      children: [
-                        Checkbox(value: false, onChanged: (value) {}),
-                        const SizedBox(width: 5),
-                        Text(
-                          "Mingguan",
-                          style: textApp.smallLabel.copyWith(
-                            color: ColorApp().dark1,
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                  SizedBox(
-                    height: 30,
-                    child: Row(
-                      children: [
-                        Checkbox(value: false, onChanged: (value) {}),
-                        const SizedBox(width: 5),
-                        Text(
-                          "Bulanan",
-                          style: textApp.smallLabel.copyWith(
-                            color: ColorApp().dark1,
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                  SizedBox(
-                    height: 30,
-                    child: Row(
-                      children: [
-                        Checkbox(value: false, onChanged: (value) {}),
-                        const SizedBox(width: 5),
-                        Text(
-                          "Tahunan",
-                          style: textApp.smallLabel.copyWith(
-                            color: ColorApp().dark1,
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                  const Divider(),
-                  const SizedBox(height: 10),
-                  Padding(
-                    padding: const EdgeInsets.symmetric(horizontal: 16.0),
-                    child: Text("Ukuran", style: textApp.bodySmall),
-                  ),
-                  RangeSlider(
-                    min: 0,
-                    max: 100,
-                    activeColor: ColorApp().mainColor,
-                    inactiveColor: ColorApp().mainColor.withOpacity(0.5),
-                    values: _sliderValue,
-                    labels: RangeLabels(
-                      _sliderValue.start.toString(),
-                      _sliderValue.end.toString(),
-                    ),
-                    onChanged: (newValues) {
-                      setState(() {
-                        _sliderValue = newValues;
-                      });
-                    },
-                  ),
-                  Padding(
-                    padding: const EdgeInsets.symmetric(horizontal: 16.0),
-                    child: Row(
-                      children: [
-                        Column(
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: [
-                            Text(
-                              "Min. size",
-                              style: textApp.extraSmallLabel.copyWith(
-                                color: ColorApp().dark1,
-                              ),
-                            ),
-                            SizedBox(
-                              height: 30,
-                              width: 80,
-                              child: TextField(
-                                cursorColor: ColorApp().mainColor,
-                                keyboardType: TextInputType.number,
-                                decoration: InputDecoration(
-                                  focusedBorder: OutlineInputBorder(
-                                      borderRadius: BorderRadius.circular(5),
-                                      borderSide: BorderSide(
-                                        color: ColorApp().mainColor,
-                                      )),
-                                  contentPadding: const EdgeInsets.only(
-                                    bottom: 5.0,
-                                    left: 5.0,
-                                  ),
-                                  border: OutlineInputBorder(
-                                    borderRadius: BorderRadius.circular(5),
-                                    borderSide: BorderSide(
-                                      color: ColorApp().mainColor,
-                                    ),
-                                  ),
-                                ),
-                              ),
-                            ),
-                          ],
-                        ),
-                        const Spacer(),
-                        Column(
-                          crossAxisAlignment: CrossAxisAlignment.end,
-                          children: [
-                            Text(
-                              "Max. size",
-                              style: textApp.extraSmallLabel.copyWith(
-                                color: ColorApp().dark1,
-                              ),
-                            ),
-                            SizedBox(
-                              height: 30,
-                              width: 80,
-                              child: TextField(
-                                cursorColor: ColorApp().mainColor,
-                                keyboardType: TextInputType.number,
-                                decoration: InputDecoration(
-                                  focusedBorder: OutlineInputBorder(
-                                      borderRadius: BorderRadius.circular(5),
-                                      borderSide: BorderSide(
-                                        color: ColorApp().mainColor,
-                                      )),
-                                  contentPadding: const EdgeInsets.only(
-                                    bottom: 5.0,
-                                    left: 5.0,
-                                  ),
-                                  border: OutlineInputBorder(
-                                    borderRadius: BorderRadius.circular(5),
-                                    borderSide: BorderSide(
-                                      color: ColorApp().mainColor,
-                                    ),
-                                  ),
-                                ),
-                              ),
-                            ),
-                          ],
-                        ),
-                      ],
-                    ),
-                  ),
-                  const Spacer(),
-                  Padding(
-                    padding: const EdgeInsets.all(16.0),
-                    child: Row(
-                      children: [
-                        ElevatedButton(
-                          style: ElevatedButton.styleFrom(
-                            backgroundColor: ColorApp().stateError,
-                            elevation: 0,
-                            shape: RoundedRectangleBorder(
-                              borderRadius: BorderRadius.circular(8),
-                            ),
-                          ),
-                          onPressed: () {},
-                          child: Text(
-                            "Delete",
-                            style: textApp.bodySmall.copyWith(
-                              color: ColorApp().light4,
-                            ),
-                          ),
-                        ),
-                        const Spacer(),
-                        SizedBox(
-                          height: 40,
-                          width: 250,
-                          child: ElevatedButton(
-                            style: ElevatedButton.styleFrom(
-                              backgroundColor: ColorApp().mainColor,
-                              elevation: 0,
-                              shape: RoundedRectangleBorder(
-                                borderRadius: BorderRadius.circular(8),
-                              ),
-                            ),
-                            onPressed: () {},
-                            child: Text(
-                              "Search",
-                              style: textApp.bodySmall.copyWith(
-                                color: ColorApp().light4,
-                              ),
-                            ),
-                          ),
-                        ),
-                      ],
-                    ),
-                  )
-                ],
-              ),
-            );
-          },
-        );
-      },
     );
   }
 }
@@ -650,13 +250,13 @@ class RecommendedCardWidget extends StatelessWidget {
                     children: [
                       Text(
                         "Warehouse Abadi",
-                        style: TextCollection.bodyNormal
+                        style: textApp.bodyNormal
                             .copyWith(color: ColorApp().dark1, fontSize: 16),
                       ),
                       const SizedBox(height: 8),
                       Text(
                         "Jakarta Barat",
-                        style: TextCollection.bodySmall.copyWith(
+                        style: textApp.bodySmall.copyWith(
                           color: ColorApp().dark1,
                           fontWeight: FontWeight.normal,
                           fontSize: 12,
@@ -665,7 +265,7 @@ class RecommendedCardWidget extends StatelessWidget {
                       const SizedBox(height: 15),
                       Text(
                         "RP.10 Jt/bln",
-                        style: TextCollection.bodyNormal.copyWith(
+                        style: textApp.bodyNormal.copyWith(
                           color: ColorApp().secondaryColor,
                         ),
                       ),
@@ -711,13 +311,13 @@ class RecommendedCardWidget extends StatelessWidget {
                     children: [
                       Text(
                         "Warehouse Abadi",
-                        style: TextCollection.bodyNormal
+                        style: textApp.bodyNormal
                             .copyWith(color: ColorApp().dark1, fontSize: 16),
                       ),
                       const SizedBox(height: 8),
                       Text(
                         "Jakarta Barat",
-                        style: TextCollection.bodySmall.copyWith(
+                        style: textApp.bodySmall.copyWith(
                           color: ColorApp().dark1,
                           fontWeight: FontWeight.normal,
                           fontSize: 12,
@@ -726,7 +326,7 @@ class RecommendedCardWidget extends StatelessWidget {
                       const SizedBox(height: 15),
                       Text(
                         "RP.10 Jt/bln",
-                        style: TextCollection.bodyNormal.copyWith(
+                        style: textApp.bodyNormal.copyWith(
                           color: ColorApp().secondaryColor,
                         ),
                       ),
@@ -772,13 +372,13 @@ class RecommendedCardWidget extends StatelessWidget {
                     children: [
                       Text(
                         "Warehouse Abadi",
-                        style: TextCollection.bodyNormal
+                        style: textApp.bodyNormal
                             .copyWith(color: ColorApp().dark1, fontSize: 16),
                       ),
                       const SizedBox(height: 8),
                       Text(
                         "Jakarta Barat",
-                        style: TextCollection.bodySmall.copyWith(
+                        style: textApp.bodySmall.copyWith(
                           color: ColorApp().dark1,
                           fontWeight: FontWeight.normal,
                           fontSize: 12,
@@ -787,7 +387,7 @@ class RecommendedCardWidget extends StatelessWidget {
                       const SizedBox(height: 15),
                       Text(
                         "RP.10 Jt/bln",
-                        style: TextCollection.bodyNormal.copyWith(
+                        style: textApp.bodyNormal.copyWith(
                           color: ColorApp().secondaryColor,
                         ),
                       ),
@@ -833,13 +433,13 @@ class RecommendedCardWidget extends StatelessWidget {
                     children: [
                       Text(
                         "Warehouse Abadi",
-                        style: TextCollection.bodyNormal
+                        style: textApp.bodyNormal
                             .copyWith(color: ColorApp().dark1, fontSize: 16),
                       ),
                       const SizedBox(height: 8),
                       Text(
                         "Jakarta Barat",
-                        style: TextCollection.bodySmall.copyWith(
+                        style: textApp.bodySmall.copyWith(
                           color: ColorApp().dark1,
                           fontWeight: FontWeight.normal,
                           fontSize: 12,
@@ -848,7 +448,7 @@ class RecommendedCardWidget extends StatelessWidget {
                       const SizedBox(height: 15),
                       Text(
                         "RP.10 Jt/bln",
-                        style: TextCollection.bodyNormal.copyWith(
+                        style: textApp.bodyNormal.copyWith(
                           color: ColorApp().secondaryColor,
                         ),
                       ),
@@ -894,13 +494,13 @@ class RecommendedCardWidget extends StatelessWidget {
                     children: [
                       Text(
                         "Warehouse Abadi",
-                        style: TextCollection.bodyNormal
+                        style: textApp.bodyNormal
                             .copyWith(color: ColorApp().dark1, fontSize: 16),
                       ),
                       const SizedBox(height: 8),
                       Text(
                         "Jakarta Barat",
-                        style: TextCollection.bodySmall.copyWith(
+                        style: textApp.bodySmall.copyWith(
                           color: ColorApp().dark1,
                           fontWeight: FontWeight.normal,
                           fontSize: 12,
@@ -909,7 +509,7 @@ class RecommendedCardWidget extends StatelessWidget {
                       const SizedBox(height: 15),
                       Text(
                         "RP.10 Jt/bln",
-                        style: TextCollection.bodyNormal.copyWith(
+                        style: textApp.bodyNormal.copyWith(
                           color: ColorApp().secondaryColor,
                         ),
                       ),
@@ -955,13 +555,13 @@ class RecommendedCardWidget extends StatelessWidget {
                     children: [
                       Text(
                         "Warehouse Abadi",
-                        style: TextCollection.bodyNormal
+                        style: textApp.bodyNormal
                             .copyWith(color: ColorApp().dark1, fontSize: 16),
                       ),
                       const SizedBox(height: 8),
                       Text(
                         "Jakarta Barat",
-                        style: TextCollection.bodySmall.copyWith(
+                        style: textApp.bodySmall.copyWith(
                           color: ColorApp().dark1,
                           fontWeight: FontWeight.normal,
                           fontSize: 12,
@@ -970,7 +570,7 @@ class RecommendedCardWidget extends StatelessWidget {
                       const SizedBox(height: 15),
                       Text(
                         "RP.10 Jt/bln",
-                        style: TextCollection.bodyNormal.copyWith(
+                        style: textApp.bodyNormal.copyWith(
                           color: ColorApp().secondaryColor,
                         ),
                       ),
@@ -988,7 +588,6 @@ class RecommendedCardWidget extends StatelessWidget {
 }
 
 class BannerWidget extends StatelessWidget {
-  //ini nanti bisa dipindah ke folder widget/component biar ga panjang" main file nya
   BannerWidget({
     super.key,
   });
